[MSFT] Add ExternEntityOp to model external entities in PlacementDB. (#2484)
The ExternEntityOp is designed to model something conceptually similar
to an external module, but it is intentionally opaque. It could be an
external instance, or anything else we might want to reserve placement
locations for. It has a symbol name, and otherwise stores an arbitrary
attribute to contain any other metadata.

These ops can be placed by creating a GlobalRefOp that refers to their
symbol with a FlatSymbolRefAttr. The PlacementDB treats them just like
any other op, but in ExportQuartusTcl, we explicitly ignore them.