[SCFToCalyx] Support either padding or slicing memory address.
In RewriteMemoryAccesses, update the logic to either pad or slice the
mismatched source signal to fit the memory interface. Previously, it
was assumed the mismatch always needs to be sliced, but this is only
the case when the signal is too wide. It can also be too small, so
support that case.