[FIRRTL][Memory] Create unmasked memory for 1 bit mask (#2482)
This commit updates the memory lowering to infer unmasked memories. 
Mask ports are required to selectively update bits of the memory data.
Masked memory are generated after flattening aggregate type memory. 
Where each mask bit can guard update to a field of the aggregate data type.
Single bit mask ports can be trivially removed by setting the enable port accordingly.
Additionally, if the mask is connected to a multi-bit constant 1, replace it with a 
single bit mask.

In this Commit: 
1.  Add an analysis in `InferReadWrite` pass to check for constant masks, and 
   replace multi-bit constant 1 bit masks with a single bit mask.
2.  Remove single bit mask ports in `LowerToHW`.
3. Update the enable signal with the mask bit.
4. Fix the memory module declaration accordingly.
5. Handle unmasked memory, in the simulation model generation in `HWMemSimImpl` 
6. Fix Lit tests.