[FIRRTL][LowerAnnos] Move annotations from instance to module
This moves instance annotations from instance ops to modules.  To do
this, we turn these annotations into non-local versions.  Theoretically,
if there is only one instance of a module, the annotation could remain
local.

This introduces proper `inner_ref` symbols on named entities for use in
NLA operations by utilization some of the new functionality in the
AnnoTarget class.  To do this, we had to start tracking a module
namespace for each module.

`ApplyState` is no longer copy-able, due to the use of unique_ptrs. This
means we have to pass `ApplyState` by reference everywhere, which is
desirable anyways.

This also fixes a bug where modules themselves could not be targeted by
annotations, since we were always creating a `hw::InnerRef` instead of a
`FlatSymbolRef` for the final element of the NLA.