[SCFToCalyx] Support parallel loop iter arg register initialization. (#2537)
Previously, all iter arg registers were assigned in a single
group. This prevents Calyx from being able to infer static timing for
this group. While we could enhance the downstream analysis, this patch
makes the potential parallelism explicit in the IR when multiple
registers are initialized. With this, Calyx is able to infer static
timing for the initialization group.